pyproject.toml
[tool.poetry]
name = "microservices"
version = "0.1.0"
description = "Microservices pet-project for вdating apps"
authors = ["danil <kotikov.lil@bkl.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.4"
werkzeug = "^3.0.3"
aiohttp = "^3.9.5"
aiodns = "^3.2.0"
redis = "^5.0.7"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
